+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/calendar/CalendarPanel.java
@@ -107,62 +107,62 @@ public final class CalendarPanel extends JPanel {
 		 * Select the previous year.<br>
 		 * Default key stroke: CTRL-DOWN ARROW
 		 */
-		public static final ControlKey<CommandControl> PREVIOUS_YEAR = CommandControl.key(keyStroke(VK_DOWN, CTRL_DOWN_MASK));
+		public static final ControlKey<CommandControl> PREVIOUS_YEAR = CommandControl.key("previousYear", keyStroke(VK_DOWN, CTRL_DOWN_MASK));
 		/**
 		 * Select the next year.<br>
 		 * Default key stroke: CTRL-UP ARROW
 		 */
-		public static final ControlKey<CommandControl> NEXT_YEAR = CommandControl.key(keyStroke(VK_UP, CTRL_DOWN_MASK));
+		public static final ControlKey<CommandControl> NEXT_YEAR = CommandControl.key("nextYear", keyStroke(VK_UP, CTRL_DOWN_MASK));
 		/**
 		 * Select the previous month.<br>
 		 * Default key stroke: SHIFT-DOWN ARROW
 		 */
-		public static final ControlKey<CommandControl> PREVIOUS_MONTH = CommandControl.key(keyStroke(VK_DOWN, SHIFT_DOWN_MASK));
+		public static final ControlKey<CommandControl> PREVIOUS_MONTH = CommandControl.key("previousMonth", keyStroke(VK_DOWN, SHIFT_DOWN_MASK));
 		/**
 		 * Select the next month.<br>
 		 * Default key stroke: SHIFT-UP ARROW
 		 */
-		public static final ControlKey<CommandControl> NEXT_MONTH = CommandControl.key(keyStroke(VK_UP, SHIFT_DOWN_MASK));
+		public static final ControlKey<CommandControl> NEXT_MONTH = CommandControl.key("nextMonth", keyStroke(VK_UP, SHIFT_DOWN_MASK));
 		/**
 		 * Select the previous week.<br>
 		 * Default key stroke: ALT-UP ARROW
 		 */
-		public static final ControlKey<CommandControl> PREVIOUS_WEEK = CommandControl.key(keyStroke(VK_UP, ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> PREVIOUS_WEEK = CommandControl.key("previousWeek", keyStroke(VK_UP, ALT_DOWN_MASK));
 		/**
 		 * Select the next week.<br>
 		 * Default key stroke: ALT-DOWN ARROW
 		 */
-		public static final ControlKey<CommandControl> NEXT_WEEK = CommandControl.key(keyStroke(VK_DOWN, ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> NEXT_WEEK = CommandControl.key("nextWeek", keyStroke(VK_DOWN, ALT_DOWN_MASK));
 		/**
 		 * Select the previous day.<br>
 		 * Default key stroke: ALT-LEFT ARROW
 		 */
-		public static final ControlKey<CommandControl> PREVIOUS_DAY = CommandControl.key(keyStroke(VK_LEFT, ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> PREVIOUS_DAY = CommandControl.key("previousDay", keyStroke(VK_LEFT, ALT_DOWN_MASK));
 		/**
 		 * Select the next day.<br>
 		 * Default key stroke: ALT-RIGHT ARROW
 		 */
-		public static final ControlKey<CommandControl> NEXT_DAY = CommandControl.key(keyStroke(VK_RIGHT, ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> NEXT_DAY = CommandControl.key("nextDay", keyStroke(VK_RIGHT, ALT_DOWN_MASK));
 		/**
 		 * Select the previous hour.<br>
 		 * Default key stroke: SHIFT-ALT-DOWN ARROW
 		 */
-		public static final ControlKey<CommandControl> PREVIOUS_HOUR = CommandControl.key(keyStroke(VK_DOWN, SHIFT_DOWN_MASK | ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> PREVIOUS_HOUR = CommandControl.key("previousHour", keyStroke(VK_DOWN, SHIFT_DOWN_MASK | ALT_DOWN_MASK));
 		/**
 		 * Select the next hour.<br>
 		 * Default key stroke: SHIFT-ALT-UP ARROW
 		 */
-		public static final ControlKey<CommandControl> NEXT_HOUR = CommandControl.key(keyStroke(VK_UP, SHIFT_DOWN_MASK | ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> NEXT_HOUR = CommandControl.key("nextHour", keyStroke(VK_UP, SHIFT_DOWN_MASK | ALT_DOWN_MASK));
 		/**
 		 * Select the previous minute.<br>
 		 * Default key stroke: CTRL-ALT-DOWN ARROW
 		 */
-		public static final ControlKey<CommandControl> PREVIOUS_MINUTE = CommandControl.key(keyStroke(VK_DOWN, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> PREVIOUS_MINUTE = CommandControl.key("previousMinute", keyStroke(VK_DOWN, CTRL_DOWN_MASK | ALT_DOWN_MASK));
 		/**
 		 * Select the next minute.<br>
 		 * Default key stroke: CTRL-ALT-UP ARROW
 		 */
-		public static final ControlKey<CommandControl> NEXT_MINUTE = CommandControl.key(keyStroke(VK_UP, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> NEXT_MINUTE = CommandControl.key("nextMinute", keyStroke(VK_UP, CTRL_DOWN_MASK | ALT_DOWN_MASK));
 
 		private ControlKeys() {}
 	}