+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTablePanel.java
@@ -200,159 +200,159 @@ public class EntityTablePanel extends JPanel {
 		 * @see EntityTablePanel#EntityTablePanel(SwingEntityTableModel, EntityEditPanel)
 		 * @see EntityTablePanel#EntityTablePanel(SwingEntityTableModel, EntityEditPanel, Consumer)
 		 */
-		public static final ControlKey<CommandControl> ADD = CommandControl.key(keyStroke(VK_INSERT));
+		public static final ControlKey<CommandControl> ADD = CommandControl.key("add", keyStroke(VK_INSERT));
 		/**
 		 * Edit the selected entity instance.<br>
 		 * Default key stroke: CTRL-INSERT
 		 * @see EntityTablePanel#EntityTablePanel(SwingEntityTableModel, EntityEditPanel)
 		 * @see EntityTablePanel#EntityTablePanel(SwingEntityTableModel, EntityEditPanel, Consumer)
 		 */
-		public static final ControlKey<CommandControl> EDIT = CommandControl.key(keyStroke(VK_INSERT, CTRL_DOWN_MASK));
+		public static final ControlKey<CommandControl> EDIT = CommandControl.key("edit", keyStroke(VK_INSERT, CTRL_DOWN_MASK));
 		/**
 		 * Select and edit a single attribute value for the selected entity instances.<br>
 		 * Default key stroke: SHIFT-INSERT
 		 * @see Config#editAttributeSelection(EditAttributeSelection)
 		 */
-		public static final ControlKey<CommandControl> EDIT_SELECTED_ATTRIBUTE = CommandControl.key(keyStroke(VK_INSERT, SHIFT_DOWN_MASK));
+		public static final ControlKey<CommandControl> EDIT_SELECTED_ATTRIBUTE = CommandControl.key("editSelectedAttribute", keyStroke(VK_INSERT, SHIFT_DOWN_MASK));
 		/**
 		 * Requests focus for the table.<br>
 		 * Default key stroke: CTRL-T
 		 */
-		public static final ControlKey<CommandControl> REQUEST_TABLE_FOCUS = CommandControl.key(keyStroke(VK_T, CTRL_DOWN_MASK));
+		public static final ControlKey<CommandControl> REQUEST_TABLE_FOCUS = CommandControl.key("requestTableFocus", keyStroke(VK_T, CTRL_DOWN_MASK));
 		/**
 		 * Toggles the condition panel between hidden, visible and advanced.<br>
 		 * Default key stroke: CTRL-ALT-S
 		 * @see TableConditionPanel#state()
 		 */
-		public static final ControlKey<CommandControl> TOGGLE_CONDITION_PANEL = CommandControl.key(keyStroke(VK_S, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> TOGGLE_CONDITION_PANEL = CommandControl.key("toggleConditionPanel", keyStroke(VK_S, CTRL_DOWN_MASK | ALT_DOWN_MASK));
 		/**
 		 * Displays a dialog for selecting a column condition panel.<br>
 		 * Default key stroke: CTRL-S
 		 */
-		public static final ControlKey<CommandControl> SELECT_CONDITION_PANEL = CommandControl.key(keyStroke(VK_S, CTRL_DOWN_MASK));
+		public static final ControlKey<CommandControl> SELECT_CONDITION_PANEL = CommandControl.key("selectConditionPanel", keyStroke(VK_S, CTRL_DOWN_MASK));
 		/**
 		 * Toggles the filter panel between hidden, visible and advanced.<br>
 		 * Default key stroke: CTRL-ALT-F
 		 * @see TableConditionPanel#state()
 		 */
-		public static final ControlKey<CommandControl> TOGGLE_FILTER_PANEL = CommandControl.key(keyStroke(VK_F, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> TOGGLE_FILTER_PANEL = CommandControl.key("toggleFilterPanel", keyStroke(VK_F, CTRL_DOWN_MASK | ALT_DOWN_MASK));
 		/**
 		 * Displays a dialog for selecting a column filter panel.<br>
 		 * Default key stroke: CTRL-SHIFT-F
 		 */
-		public static final ControlKey<CommandControl> SELECT_FILTER_PANEL = CommandControl.key(keyStroke(VK_F, CTRL_DOWN_MASK | SHIFT_DOWN_MASK));
+		public static final ControlKey<CommandControl> SELECT_FILTER_PANEL = CommandControl.key("selectFilterPanel", keyStroke(VK_F, CTRL_DOWN_MASK | SHIFT_DOWN_MASK));
 		/**
 		 * Moves the selection up.<br>
 		 * Default key stroke: ALT-SHIFT-UP
 		 */
-		public static final ControlKey<CommandControl> MOVE_SELECTION_UP = CommandControl.key(keyStroke(VK_UP, ALT_DOWN_MASK | SHIFT_DOWN_MASK));
+		public static final ControlKey<CommandControl> MOVE_SELECTION_UP = CommandControl.key("moveSelectionUp", keyStroke(VK_UP, ALT_DOWN_MASK | SHIFT_DOWN_MASK));
 		/**
 		 * Moves the selection down.<br>
 		 * Default key stroke: ALT-SHIFT-DOWN
 		 */
-		public static final ControlKey<CommandControl> MOVE_SELECTION_DOWN = CommandControl.key(keyStroke(VK_DOWN, ALT_DOWN_MASK | SHIFT_DOWN_MASK));
+		public static final ControlKey<CommandControl> MOVE_SELECTION_DOWN = CommandControl.key("moveSelectionDown", keyStroke(VK_DOWN, ALT_DOWN_MASK | SHIFT_DOWN_MASK));
 		/**
 		 * The main print action<br>
 		 * Default key stroke: CTRL-P
 		 */
-		public static final ControlKey<CommandControl> PRINT = CommandControl.key(keyStroke(VK_P, CTRL_DOWN_MASK));
+		public static final ControlKey<CommandControl> PRINT = CommandControl.key("print", keyStroke(VK_P, CTRL_DOWN_MASK));
 		/**
 		 * Triggers the {@link ControlKeys#DELETE} control.<br>
 		 * Default key stroke: DELETE
 		 */
-		public static final ControlKey<CommandControl> DELETE = CommandControl.key(keyStroke(VK_DELETE));
+		public static final ControlKey<CommandControl> DELETE = CommandControl.key("delete", keyStroke(VK_DELETE));
 		/**
 		 * Displays the table popup menu, if one is available.<br>
 		 * Default key stroke: CTRL-G
 		 */
-		public static final ControlKey<CommandControl> DISPLAY_POPUP_MENU = CommandControl.key(keyStroke(VK_G, CTRL_DOWN_MASK));
+		public static final ControlKey<CommandControl> DISPLAY_POPUP_MENU = CommandControl.key("displayPopupMenu", keyStroke(VK_G, CTRL_DOWN_MASK));
 		/**
 		 * Displays the entity menu, if one is available.<br>
 		 * Default key stroke: CTRL-ALT-V
 		 */
-		public static final ControlKey<CommandControl> DISPLAY_ENTITY_MENU = CommandControl.key(keyStroke(VK_V, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> DISPLAY_ENTITY_MENU = CommandControl.key("displayEntityMenu", keyStroke(VK_V, CTRL_DOWN_MASK | ALT_DOWN_MASK));
 		/**
 		 * A {@link Controls} instance containing controls for printing.
 		 */
-		public static final ControlKey<Controls> PRINT_CONTROLS = Controls.key();
+		public static final ControlKey<Controls> PRINT_CONTROLS = Controls.key("printControls");
 		/**
 		 * A {@link Controls} instance containing any additional popup menu controls.
 		 * @see #addPopupMenuControls(Controls)
 		 */
-		public static final ControlKey<Controls> ADDITIONAL_POPUP_MENU_CONTROLS = Controls.key();
+		public static final ControlKey<Controls> ADDITIONAL_POPUP_MENU_CONTROLS = Controls.key("additionalPopupMenuControls");
 		/**
 		 * A {@link Controls} instance containing any additional toolbar controls.
 		 * @see #addToolBarControls(Controls)
 		 */
-		public static final ControlKey<Controls> ADDITIONAL_TOOLBAR_CONTROLS = Controls.key();
+		public static final ControlKey<Controls> ADDITIONAL_TOOLBAR_CONTROLS = Controls.key("additionalToolBarControls");
 		/**
 		 * A {@link Control} for viewing the dependencies of the selected entities.
 		 */
-		public static final ControlKey<CommandControl> VIEW_DEPENDENCIES = CommandControl.key();
+		public static final ControlKey<CommandControl> VIEW_DEPENDENCIES = CommandControl.key("viewDependencies");
 		/**
 		 * A {@link Controls} instance containing edit controls for all editable attributes.
 		 * @see Config#editAttributeSelection(EditAttributeSelection)
 		 */
-		public static final ControlKey<Controls> EDIT_ATTRIBUTE_CONTROLS = Controls.key();
+		public static final ControlKey<Controls> EDIT_ATTRIBUTE_CONTROLS = Controls.key("editAttributeControls");
 		/**
 		 * A {@link Control} for displaying a dialog for selecting the visible table columns.
 		 * @see Config#columnSelection(ColumnSelection)
 		 */
-		public static final ControlKey<CommandControl> SELECT_COLUMNS = CommandControl.key();
+		public static final ControlKey<CommandControl> SELECT_COLUMNS = CommandControl.key("selectColumns");
 		/**
 		 * A {@link Controls} instance containing a {@link ToggleControl} for each columns visibility.
 		 * @see Config#columnSelection(ColumnSelection)
 		 */
-		public static final ControlKey<Controls> TOGGLE_COLUMN_CONTROLS = Controls.key();
+		public static final ControlKey<Controls> TOGGLE_COLUMN_CONTROLS = Controls.key("toggleColumnControls");
 		/**
 		 * A {@link Control} for resetting the columns to their original visibility and location.
 		 */
-		public static final ControlKey<CommandControl> RESET_COLUMNS = CommandControl.key();
+		public static final ControlKey<CommandControl> RESET_COLUMNS = CommandControl.key("resetColumns");
 		/**
 		 * A {@link Control} for displaying a dialog for configuring the column auto-resize-mode.
 		 * @see JTable#setAutoResizeMode(int)
 		 */
-		public static final ControlKey<CommandControl> COLUMN_AUTO_RESIZE_MODE = CommandControl.key();
+		public static final ControlKey<CommandControl> COLUMN_AUTO_RESIZE_MODE = CommandControl.key("columnAutoResizeMode");
 		/**
 		 * A {@link Control} for toggling between single and multi selection mode.
 		 */
-		public static final ControlKey<ToggleControl> SELECTION_MODE = ToggleControl.key();
+		public static final ControlKey<ToggleControl> SELECTION_MODE = ToggleControl.key("selectionMode");
 		/**
 		 * A {@link Control} for clearing the data from the table.
 		 * @see SwingEntityTableModel#clear()
 		 */
-		public static final ControlKey<CommandControl> CLEAR = CommandControl.key();
+		public static final ControlKey<CommandControl> CLEAR = CommandControl.key("clear");
 		/**
 		 * A {@link Control} for refreshing the table data.<br>
 		 * Default key stroke: ALT-R
 		 * @see SwingEntityTableModel#refresh()
 		 */
-		public static final ControlKey<CommandControl> REFRESH = CommandControl.key(keyStroke(VK_R, ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> REFRESH = CommandControl.key("refresh", keyStroke(VK_R, ALT_DOWN_MASK));
 		/**
 		 * A {@link ToggleControl} for showing/hiding the summary panel.
 		 */
-		public static final ControlKey<ToggleControl> TOGGLE_SUMMARY_PANEL = ToggleControl.key();
+		public static final ControlKey<ToggleControl> TOGGLE_SUMMARY_PANEL = ToggleControl.key("toggleSummaryPanel");
 		/**
 		 * A {@link Controls} instance containing the condition panel controls.
 		 */
-		public static final ControlKey<Controls> CONDITION_CONTROLS = Controls.key();
+		public static final ControlKey<Controls> CONDITION_CONTROLS = Controls.key("conditionControls");
 		/**
 		 * A {@link Controls} instance containing the filter panel controls.
 		 */
-		public static final ControlKey<Controls> FILTER_CONTROLS = Controls.key();
+		public static final ControlKey<Controls> FILTER_CONTROLS = Controls.key("filterControls");
 		/**
 		 * A {@link Control} for clearing the table selection.
 		 */
-		public static final ControlKey<CommandControl> CLEAR_SELECTION = CommandControl.key();
+		public static final ControlKey<CommandControl> CLEAR_SELECTION = CommandControl.key("clearSelection");
 		/**
 		 * A {@link Control} for copying the selected cell data.<br>
 		 * Default key stroke: CTRL-ALT-C
 		 */
-		public static final ControlKey<CommandControl> COPY_CELL = CommandControl.key(keyStroke(VK_C, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+		public static final ControlKey<CommandControl> COPY_CELL = CommandControl.key("copyCell", keyStroke(VK_C, CTRL_DOWN_MASK | ALT_DOWN_MASK));
 		/**
 		 * A {@link Control} for copying the table rows with header.
 		 */
-		public static final ControlKey<CommandControl> COPY_ROWS = CommandControl.key();
+		public static final ControlKey<CommandControl> COPY_ROWS = CommandControl.key("copyRows");
 		/**
 		 * A {@link Controls} instance containing controls for copying either cell or table data.
 		 * <li>{@link ControlKeys#COPY_ROWS ControlKeys#COPY_ROWS}</li>
@@ -360,7 +360,7 @@ public class EntityTablePanel extends JPanel {
 		 * @see #COPY_CELL
 		 * @see #COPY_ROWS
 		 */
-		public static final ControlsKey COPY_CONTROLS = Controls.key(Controls.layout(asList(COPY_CELL, COPY_ROWS)));
+		public static final ControlsKey COPY_CONTROLS = Controls.key("copyControls", Controls.layout(asList(COPY_CELL, COPY_ROWS)));
 		/**
 		 * A {@link Controls} instance containing controls for configuring columns.
 		 * <li>{@link ControlKeys#SELECT_COLUMNS ControlKeys#SELECT_COLUMNS} or {@link ControlKeys#TOGGLE_COLUMN_CONTROLS ControlKeys#TOGGLE_COLUMN_CONTROLS}</li>
@@ -371,12 +371,12 @@ public class EntityTablePanel extends JPanel {
 		 * @see #RESET_COLUMNS
 		 * @see #COLUMN_AUTO_RESIZE_MODE
 		 */
-		public static final ControlsKey COLUMN_CONTROLS = Controls.key(Controls.layout(asList(SELECT_COLUMNS, TOGGLE_COLUMN_CONTROLS, RESET_COLUMNS, COLUMN_AUTO_RESIZE_MODE)));
+		public static final ControlsKey COLUMN_CONTROLS = Controls.key("columnControls", Controls.layout(asList(SELECT_COLUMNS, TOGGLE_COLUMN_CONTROLS, RESET_COLUMNS, COLUMN_AUTO_RESIZE_MODE)));
 		/**
 		 * Requests focus for the table search field.<br>
 		 * Default key stroke: CTRL-F
 		 */
-		public static final ControlKey<CommandControl> REQUEST_SEARCH_FIELD_FOCUS = CommandControl.key(keyStroke(VK_F, CTRL_DOWN_MASK));
+		public static final ControlKey<CommandControl> REQUEST_SEARCH_FIELD_FOCUS = CommandControl.key("requestSearchFieldFocus", keyStroke(VK_F, CTRL_DOWN_MASK));
 
 		private ControlKeys() {}
 	}